backend_app/requirement.txt
altgraph==0.10.2
bdist-mpkg==0.5.0
bonjour-py==0.3
click==4.1
configobj==5.0.6
Django==1.8.3
django-allauth==0.23.0
django-cors-headers==1.1.0
django-rest-auth==0.5.0
django-tinymce==2.0.5
djangorestframework==3.2.4
djangorestframework-jwt==1.7.1
djangorestframework-recursive==0.1.1
drf-nested-routers==0.10.0
google-api-python-client==1.4.2
gunicorn==19.3.0
httplib2==0.9.2
macholib==1.5.1
matplotlib==1.3.1
modulegraph==0.10.4
numpy==1.8.0rc1
oauth2client==1.5.1
oauthlib==1.0.3
pbr==1.3.0
prompt-toolkit==0.42
py2app==0.7.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
Pygments==2.0.2
PyJWT==1.4.0
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
python-openid==2.2.5
pytz==2013.7
requests==2.8.1
requests-oauthlib==0.5.0
rsa==3.2.3
scipy==0.13.0b1
simplejson==3.8.1
six==1.10.0
sqlparse==0.1.14
stevedore==1.6.0
Twisted==13.2.0
uritemplate==0.6
virtualenv==13.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.6.0
wcwidth==0.1.4
xattr==0.6.4
zope.interface==4.1.1
